requirements.txt
flask==2.3.3
python-dotenv==1.0.0
supabase==2.0.3
flask-wtf==1.2.1
wtforms==3.0.1
gunicorn==21.2.0
pyjwt==2.8.0
flask-cors==4.0.0
requests==2.31.0